+++ NEWS.md
@@ -21,6 +21,10 @@ for density and histogram types. (#209 @zeileis)
 character variable automatically being coerced to a factor. (#219 @zeileis)
 - Respect `xlim` and `ylim` when explicitly supplied by the user. (Thanks to
 @mclements for code submission #221)
+- Axis titles for flipped (horizontal) boxplots are appropriately swapped too.
+(#223 @grantmcdermott)
+- Ribbon plots without `ymin` or `ymax` args, now inherit these values from `y`
+(#224 @grantmcdermott)
 
 Internals: